Added new package "it.cnr.iasi.leks.bedspread.impl.weights.ic"  and moved some classes into. Reference were fixed with this commit
