+++ bedspread-core/src/main/java/it/cnr/iasi/leks/bedspread/impl/weights/ic/SemanticWeighting_IC.java
@@ -16,10 +16,8 @@
  *	 You should have received a copy of the GNU General Public License
  *	 along with this source.  If not, see <http://www.gnu.org/licenses/>.
  */
-package it.cnr.iasi.leks.bedspread.impl.weights;
+package it.cnr.iasi.leks.bedspread.impl.weights.ic;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Set;
 
 import it.cnr.iasi.leks.bedspread.Node;
